rename `ColorPaletteCard` to `ColorCard`
